Homework 19/20
During lesson 20 I had already started making changes to 19, so I renamed homework 19 branch and file to Homework 19_20 since I sort of messed that up, I guess.
