+++ build.gradle
@@ -20,8 +20,8 @@ test {
         suites "src/test/resources/TestNG.xml"
     }
 }
-task testHomework19(type: Test) {
+task testHomework19_20(type: Test) {
     useTestNG() {
-        suites "src/test/resources/Homework19.xml"
+        suites "src/test/resources/Homework19_20.xml"
     }
 }
