+++ src/test/java/BaseTest.java
@@ -4,14 +4,25 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.annotations.*;
-
 import java.time.Duration;
+import org.openqa.selenium.support.ui.WebDriverWait;
+
+
 
 
 public class BaseTest {
 
-    public static WebDriver driver = null;
+    public static WebDriver driver;
+    WebDriverWait wait;
+
+//    These aren't working. For now, passing the locators in with the methods.
+//    By emailField = By.cssSelector("input[type='email']");
+//    By passwordField = By.cssSelector("input[type='password']");
+//    By submitBtn = By.cssSelector("button[type ='submit']");
+
+
 
     public static String url = "https://qa.koel.app/";
 
@@ -28,9 +39,10 @@ public class BaseTest {
         options.addArguments("--disable-notifications");
 
         driver = new ChromeDriver(options);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+        wait = new WebDriverWait(driver, Duration.ofSeconds(4));
+//        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
         driver.manage().window().maximize();
-        url = BaseUrl;
+//        url = BaseUrl;
         navigateToPage();
     }
 
@@ -42,18 +54,26 @@ public class BaseTest {
         driver.get(url);
 
     }
-    public static void provideEmail(String email) {
-        WebElement emailField = driver.findElement(By.cssSelector("input[type='email']"));
-        emailField.clear();
-        emailField.sendKeys(email);
+    public void provideEmail(String email) {
+        WebElement emailElement = wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector("input[type='email']")));
+//        Passing emailField isn't working, so not using it for now
+//        WebElement emailElement = wait.until(ExpectedConditions.elementToBeClickable(By.cssSelector("emailField")));
+//        WebElement emailField = driver.findElement(By.cssSelector("input[type='email']"));
+        emailElement.clear();
+        emailElement.sendKeys(email);
     }
-    public static void providePassword(String password) {
-        WebElement passwordField = driver.findElement(By.cssSelector("input[type='password']"));
-        passwordField.clear();
-        passwordField.sendKeys(password);
+    public void providePassword(String password) {
+        WebElement passwordElement = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("input[type='password']")));
+//        Passing passWordField isn't working, so not using it for now
+//        WebElement passwordElement = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("passwordField")));
+//        WebElement passwordField = driver.findElement(By.cssSelector("input[type='password']"));
+        passwordElement.clear();
+        passwordElement.sendKeys(password);
     }
-    public static void clickSubmit() {
-        WebElement submit = driver.findElement(By.cssSelector("button[type = 'submit']"));
+    public void clickSubmit() {
+        WebElement submit = wait.until(ExpectedConditions.visibilityOfElementLocated(By.cssSelector("button[type='submit']")));
+//        Passing submitBtn isn't working.
+//        WebElement submit = driver.findElement(By.cssSelector("submitBtn"));
         submit.click();
 
     }
