+++ src/test/resources/TestNG.xml
@@ -9,7 +9,7 @@
             <class name="LoginTests"/>
             <class name="BaseTest"/>
             <class name="Homework18"/>
-            <class name="Homework19"/>
+            <class name="Homework19_20"/>
 
         </classes>
     </test>
